Pathfinder: separate searching and parsing
Rather than searching for a match at every byte while parsing,
it now does a search first, using an algorithm inspired by M4.
This speeds it up significantly, both because there are fewer
search steps to do, and because there are fewer matches for the
parser to sort thorough. Compression improves on shorter chain
lengths, and decreses on higher ones.